undo some stuff in interval

+++ core/src/data/interval/primitive.rs
@@ -114,7 +114,7 @@ impl Div<i32> for Interval {
 
     fn div(self, rhs: i32) -> Self {
         match self {
-            Interval::Month(v) => Interval::Month(v / rhs as i32),
+            Interval::Month(v) => Interval::Month(v / rhs),
             Interval::Microsecond(v) => Interval::Microsecond(v / rhs as i64),
         }
     }
@@ -125,7 +125,7 @@ impl Div<i64> for Interval {
 
     fn div(self, rhs: i64) -> Self {
         match self {
-            Interval::Month(v) => Interval::Month((v as i64 / rhs) as i32),
+            Interval::Month(v) => Interval::Month(((v as i64) / rhs) as i32),
             Interval::Microsecond(v) => Interval::Microsecond((v / rhs) as i64),
         }
     }
@@ -169,7 +169,7 @@ impl Div<Interval> for i32 {
 
     fn div(self, rhs: Interval) -> Interval {
         match rhs {
-            Interval::Month(v) => Interval::Month(self / v as i32),
+            Interval::Month(v) => Interval::Month(self / v),
             Interval::Microsecond(v) => Interval::Microsecond(self as i64 / v),
         }
     }
@@ -180,7 +180,7 @@ impl Div<Interval> for i64 {
 
     fn div(self, rhs: Interval) -> Interval {
         match rhs {
-            Interval::Month(v) => Interval::Month((self / v as i64) as i32),
+            Interval::Month(v) => Interval::Month((self / (v as i64)) as i32),
             Interval::Microsecond(v) => Interval::Microsecond(self / v),
         }
     }
